+++ websocketpp/transport/iostream/endpoint.hpp
@@ -117,6 +117,24 @@ public:
         return m_is_secure;
     }
     
+    /// Sets the write handler
+    /**
+     * The write handler is called when the iostream transport receives data
+     * that needs to be written to the appropriate output location. This handler
+     * can be used in place of registering an ostream for output.
+     *
+     * The signature of the handler is 
+     * `lib::error_code (connection_hdl, char const *, size_t)` The
+     * code returned will be reported and logged by the core library.
+     *
+     * @since 0.5.0
+     *
+     * @param h The handler to call on connection shutdown.
+     */
+    void set_write_handler(write_handler h) {
+        m_write_handler = h;
+    }
+    
     /// Sets the shutdown handler
     /**
      * The shutdown handler is called when the iostream transport receives a
@@ -181,11 +199,15 @@ protected:
         if (m_shutdown_handler) {
             tcon->set_shutdown_handler(m_shutdown_handler);
         }
+        if (m_write_handler) {
+            tcon->set_write_handler(m_write_handler);
+        }
         return lib::error_code();
     }
 private:
     std::ostream *  m_output_stream;
     shutdown_handler m_shutdown_handler;
+    write_handler   m_write_handler;
     
     elog_type *     m_elog;
     alog_type *     m_alog;