+++ changelog.md
@@ -9,6 +9,9 @@ HEAD
   iostream transport. This provides a clean interface for triggering the shut 
   down of external sockets and other cleanup without hooking in to higher level
   WebSocket handlers.
+- Feature: Adds the ability to register a write handler when using the iostream
+  transport. This handler can be used to handle transport output in place of
+  registering an ostream to write to.
 - Improvement: Message payload logging now prints text for text messages rather
   than binary.
 - Documentation: Add Sending & Receiving Messages step to chapter one of the